requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
flake8==4.0.1
idna==3.3
isort==5.9.3
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.1
requests==2.26.0
urllib3==1.26.7
